Rework port scan loop to eagerly reserve a display
Instead of checking for port availability by connecting to a Socket,
attempt to open a ServerSocket on the port (with the same
failover-retry mechanism for in-use ports). The ServerSocket is then
disposed immediately before Xvfb is launched on the same port.

This has the advantage of eagerly locking the port against other checks,
and appears to help mitigate the race condition which opens multiple
Xvfb instances on the same port.