Add lockfile mechanism for ports, add shutdown handler
Further mitigate the race condition on port selection by creating a
lockfile for a successful port grab. This persists the port choice
throughout the Xvfb run and removes the narrow window of opportunity for
another process to find the same 'unused' port.

Also add a shutdown handler to xvfb:run which cleans up the process and
lockfile if xvfb:stop is not executed.